test: further split coordinator coverage tests
